Create separate package for java beans
