+++ src/main/java/com/highperformancespark/examples/objects/JavaCoffeeShop.java
@@ -1,0 +1,27 @@
+package com.highperformancespark.examples.objects;
+
+public class JavaCoffeeShop {
+  private String zip;
+  private String name;
+
+  public JavaCoffeeShop(String zip, String name) {
+    this.zip = zip;
+    this.name = name;
+  }
+
+  public String getZip() {
+    return zip;
+  }
+
+  public void setZip(String zip) {
+    this.zip = zip;
+  }
+
+  public String getName() {
+    return name;
+  }
+
+  public void setName(String name) {
+    this.name = name;
+  }
+}
